+++ R/popsNetwork.R
@@ -31,13 +31,14 @@
 #' @description
 #' Builds and displays populations network.
 #' @param obj an `IFC_data` object extracted with features extracted.
-#' @param hierarchical whether to display network using a hierarchical layout or not. Default is TRUE.
-#' @param color_mode Whether to extract colors from 'obj' in "white" or "black" mode. Default is "white".
-#' @param highlight population to permanently highlight. If found in 'obj', this population will be displayed with its color. Default is NULL.
-#' @param seed If you provide a seed manually, the layout will be the same every time. Default is NULL.
-#' @param direction The direction of the hierarchical layout. Default is 'LR'.\cr
-#' The available options are: 'UD', 'DU', 'LR', 'RL'. To simplify: up-down, down-up, left-right, right-left.
-#' @param weighted whether to scale population's node size according to count. Default is TRUE.
+#' @param hierarchical whether to display network using a hierarchical layout or not. Default is \code{TRUE}, and will be forced to \code{TRUE} when \code{'flatten'} is \code{TRUE}.
+#' @param color_mode Whether to extract colors from \code{'obj'} in \code{"white"} or \code{"black"} mode. Default is \code{"white"}, only apply when \code{'flatten'} is \code{FALSE}.
+#' @param highlight population to permanently highlight. If found in \code{'obj'}, this population will be displayed with its color. Default is \code{NULL}, only apply when \code{'flatten'} is \code{FALSE}.
+#' @param seed If you provide a seed manually, the layout will be the same every time. Default is \code{NULL}, , only apply when \code{'flatten'} is \code{FALSE}.
+#' @param direction The direction of the hierarchical layout. Default is \code{'LR'} and will be forced to \code{'UD'} when \code{'flatten'} is \code{TRUE}.\cr
+#' The available options are: \code{'UD'}, \code{'DU'}, \code{'LR'}, \code{'RL'}. To simplify: up-down, down-up, left-right, right-left.
+#' @param weighted whether to scale population's node size according to count. Default is \code{TRUE}, only apply when \code{'flatten'} is \code{FALSE}.
+#' @param flatten whether to return a flatten data.frame of stats whose row.names represent the hierarchy tree. Default is \code{FALSE}.
 #' @param ... other argument to be passed.
 #' @examples
 #' if(requireNamespace("IFCdata", quietly = TRUE)) {
@@ -50,19 +51,36 @@
 #'                   'https://gitdemont.github.io/IFCdata/',
 #'                   'to install extra files required to run this example.'))
 #' }
-#' @return a \pkg{visNetwork} object.
+#' @return a \pkg{visNetwork} object or a data.frame if \code{'flatten'} is \code{TRUE}.
 #' @export
-popsNetwork = function(obj, hierarchical=TRUE, color_mode="white", highlight=NULL, seed=NULL, direction="LR", weighted=TRUE, ...) {
+popsNetwork = function(obj, hierarchical=TRUE, color_mode="white", highlight=NULL, seed=NULL, direction="LR", weighted=TRUE, flatten=FALSE, ...) {
   dots = list(...)
   if(missing(obj)) stop("'obj' can't be missing")
   if(!("IFC_data"%in%class(obj))) stop("'obj' is not of class `IFC_data`")
   if(length(obj$pops)==0) stop("please use argument 'extract_features' = TRUE with ExtractFromDAF() or ExtractFromXIF() and ensure that features were correctly extracted")
+  flatten = as.logical(flatten); assert(flatten, len=1, alw=c(TRUE,FALSE))
+  color_mode=which(c("black","white")%in%color_mode)
+  if(flatten) {
+    if(length(obj$stats) == 0) obj$stats = get_pops_stats(obj$pops, as.integer(obj$description$ID$objcount))
+    tree = popsTree(obj$pops)
+    name = names(tree)
+    styl = popsStyle(obj)
+    d = cbind(as.data.frame(obj$stats[name, 3:5]))
+    # TODO add color and style
+    rownames(d) = sapply(seq_along(tree), simplify = FALSE, FUN=function(i) {
+      paste0(collapse="",
+        paste0(tree[[i]],collapse=""),
+        paste0(ifelse(i==1,"","\u202f"),styl[[name[i]]],"\u202f",name[i],"\u202f")
+      )
+    })
+    return(d)
+  }
   if(!is.null(seed)) {seed=na.omit(as.integer(seed)); assert(seed, len=1, typ="integer")}
   hierarchical = as.logical(hierarchical); assert(hierarchical, len=1, alw=c(TRUE,FALSE))
   weighted = as.logical(weighted); assert(weighted, len=1, alw=c(TRUE,FALSE))
-  assert(color_mode, len=1, alw=c("black","white"))
+
   assert(direction, len=1, alw=c("UD", "DU", "LR", "RL"))
-  color_mode=which(c("black","white")%in%color_mode)
+
   # determine levels for hierarchical layout
   lev = popsGetLevels(obj$pops)
   size = table(lev)
@@ -84,7 +102,11 @@ popsNetwork = function(obj, hierarchical=TRUE, color_mode="white", highlight=NUL
     if(typ=="Boolean") tit = paste0(tit, "<br>def: ", paste0(p$split,collapse=" "))
     if(typ=="Graphical") {
       tit = paste0(tit, ", %", p$base, ": <b>", round(100*sum(p$obj)/sum(obj$pops[[p$base]]$obj),2), "</b>")
-      tit = paste0(tit, "<br><b>", obj$regions[[p$region]]$type, "</b><br><b>x</b>: ", p$fx)
+      sync_typ = sync_type(obj$regions[[p$region]])
+      sync_num = sync_part(obj$regions[[p$region]], sync_typ)
+      type = obj$regions[[p$region]]$type
+      if(sync_num != "") type = paste0(type, " ", sync_typ, sync_num)
+      tit = paste0(tit, "<br><b>", type, "</b><br><b>x</b>: ", p$fx)
       if(!is.null(p$fy)) tit = paste(tit, "<br><b>y</b>: ", p$fy)
       tit = paste0(tit, "<table> <tr> <th>x</th> <th>y</th> </tr>")
       tit = paste0(tit, paste("<tr>", apply(do.call(cbind, args = obj$regions[[p$region]][c("x","y")]), 1, FUN=function(r) paste("<td>",r,"</td>",collapse=" ")), "</tr>", collapse = "\n"), "</table>")